% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COI_Proportion_Map.R
\name{coi_proportions}
\alias{coi_proportions}
\title{Complexity Of Infection Plots}
\usage{
coi_proportions(
  df,
  coi_column,
  save_output = TRUE,
  label_repel = 1.3,
  period_name = "Full",
  label_size = 2.5,
  map_data,
  donut_chart_size = 0.5,
  time = NULL,
  plot_width = 11,
  plot_height = 6,
  plot_dpi = 600,
  ...
)
}
\arguments{
\item{df}{Combined GRC data frame}

\item{coi_column}{The column containing containing the COI information.}

\item{save_output}{Logical. If \code{TRUE}, saves the plot as a JPEG file in the output directory (default: \code{FALSE}).}

\item{label_repel}{Numeric. Controls the distance of the label from the points on the map. Default: \code{1.3}.}

\item{period_name}{The period name for the plot. Defualt: \code{FULL}}

\item{label_size}{Numeric. Controls the size of location labels on the map. Default: \code{2.5}.}

\item{map_data}{A list containing the shape file and longitude-latitude data for mapping.}

\item{donut_chart_size}{Numeric. Scales the maximum pie chart size. Default: \code{0.5}}

\item{time}{Optional. A list defining time periods for filtering the data.}

\item{plot_width}{Sets the width (in inches) of the saved plot. Default: \code{11}}

\item{plot_height}{Sets the height (in inches) of the saved plot. Default: \code{6}.}

\item{plot_dpi}{Numeric. Sets the resolution (dots per inch) for the saved plot. Default: \code{600}.}

\item{...}{Additional arguments passed to other functions.}
}
\value{
A list containing:
\itemize{
\item \code{COI_SPlot}: A scatter plots showing the all samples for ech location colored by 2 categories, Single_Infection(COI = 1) and Multiple_Infections(COI>=2)
\item \code{COI_PMap}: A donut chart showing th proportions of Single_Infection to Multiple_Infections
\item \code{COI_Table}: A data frame summarizing the counts and percentages of the COI categories across locations.
}
}
\description{
This function generates plots showing the proportion of Mono genomic Infection to Polygenomic Infections,
in order words single infections to mixed infections.
}
