% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiversityMap.R
\name{create_dm}
\alias{create_dm}
\title{Internal function to create the diversity map}
\usage{
create_dm(
  df,
  snp_data,
  period_name = "Full",
  label_size = 2.5,
  map_data,
  circle_num_size = 3.1,
  scale_circle_size = 10,
  save_output = TRUE,
  ...
)
}
\arguments{
\item{df}{Final GRC data frame}

\item{label_size}{Numeric. Controls the size of location labels on the map. Default: \code{2.5}.}

\item{map_data}{A list containing the shape file and longitude-latitude data for mapping.}

\item{save_output}{Logical. If \code{TRUE}, saves the plot as a JPEG file in the output directory (default: \code{FALSE}).}
}
\description{
Internal function to create the diversity map
}
\keyword{internal}
