% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mutation_Frequency_Plots.R
\name{create_m_plots}
\alias{create_m_plots}
\title{Internal function to create the mutation plots}
\usage{
create_m_plots(
  df,
  gene,
  gene_col,
  save_output = TRUE,
  period_name = "Full",
  map_data,
  label_size = 2.5,
  circle_num_size = 3.1,
  scale_circle_size = 10,
  include_mixed = FALSE,
  ...
)
}
\arguments{
\item{df}{Final GRC data frame}

\item{gene}{The gene to analyze ("pfcrt", "pfdhps", "pfdhfr", "pfmdr1")}

\item{gene_col}{The column containing haplotype data (e.g., "PfCRT")}

\item{save_output}{Logical. If \code{TRUE}, saves the plot as a JPEG file in the output directory (default: \code{FALSE}).}

\item{period_name}{The period name for the plot. Defualt: \code{FULL}}

\item{map_data}{A list containing the shape file and longitude-latitude data for mapping.}

\item{label_size}{Numeric. Controls the size of location labels on the map. Default: \code{2.5}.}

\item{circle_num_size}{Numeric. Controls the numbers inside the circles. Default: \code{3.1}.}

\item{scale_circle_size}{Numeric. Scales the maximum circle size. Default: \code{10}.}

\item{include_mixed}{Boolean, whether to include mixed haplotypes in counts}
}
\description{
Internal function to create the mutation plots
}
\keyword{internal}
