% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Drug_Distribution_Plots.R
\name{create_plots}
\alias{create_plots}
\title{Internal Function to Create Drug Distribution Plots}
\usage{
create_plots(
  df,
  drug_col,
  period_name = "Full",
  save_output = TRUE,
  colors = c(resistant = "#525CEB", mixed_resistant = "#808000", sensitive = "#800000"),
  ...
)
}
\arguments{
\item{df}{Final GRC data frame.}

\item{drug_col}{The name of the column representing the drug conditions (e.g., "Chloroquine" with
categories like Resistant, Mixed Resistant, and Sensitive).}

\item{period_name}{The period name for the plot. Defualt: \code{FULL}}

\item{save_output}{Logical. If \code{TRUE}, saves the plot as a JPEG file in the output directory (default: \code{FALSE}).}

\item{colors}{A named vector of colors for the categories in \code{drug_col}}
}
\description{
Internal Function to Create Drug Distribution Plots
}
\keyword{internal}
