% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleCountMap.R
\name{create_sc_map}
\alias{create_sc_map}
\title{Internal Function to Create Sample Count Map}
\usage{
create_sc_map(
  df,
  map_data,
  save_output = TRUE,
  circle_num_size = 3.1,
  label_repel = 1.3,
  period_name = "Full",
  label_size = 2.5,
  scale_circle_size = 11,
  ...
)
}
\arguments{
\item{df}{Final GRC data frame.}

\item{map_data}{A list containing the shape file and longitude-latitude data for mapping.}

\item{save_output}{Logical. If \code{TRUE}, saves the plot as a JPEG file in the output directory (default: \code{FALSE}).}

\item{circle_num_size}{Numeric. Controls the numbers inside the circles. Default: \code{3.1}}

\item{label_repel}{Numeric. Controls the distance of the label from the points on the map. Default: \code{1.3}}

\item{period_name}{The period name for the plot. Default: \code{FULL}}

\item{label_size}{Numeric. Controls the size of location labels on the map. Default: \code{2.5}.}

\item{scale_circle_size}{Numeric. Scales the maximum circle size. Default: \code{11}.}

\item{...}{Additional arguments passed to other functions.}
}
\description{
Internal Function to Create Sample Count Map
}
\keyword{internal}
