% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diversity_Map.R
\name{diversity_map}
\alias{diversity_map}
\title{Create Diversity Map}
\usage{
diversity_map(
  df,
  snp_data,
  label_size = 2.5,
  map_data,
  label_repel = 1.3,
  circle_num_size = 3.1,
  scale_circle_size = 10,
  save_output = TRUE,
  plot_width = 11,
  plot_height = 8,
  plot_dpi = 600,
  ...
)
}
\arguments{
\item{df}{Combined GRC data frame}

\item{snp_data}{A data frame of SNP data, where each row represents a sample and each column corresponds to a
genetic locus (e.g., "Pf3D7_"). The row names should correspond to the "Sample Internal ID".}

\item{label_size}{Numeric. Controls the size of location labels on the map. Default: \code{2.5}.}

\item{map_data}{A list containing the shape file and longitude-latitude data for mapping.}

\item{label_repel}{Numeric. Controls the distance of the label from the points on the map. Default: \code{1.3}.}

\item{circle_num_size}{Numeric. Controls the numbers inside the circles. Default: \code{3.1}.}

\item{scale_circle_size}{Numeric. Scales the maximum circle size. Default: \code{10}.}

\item{save_output}{Logical. If \code{TRUE}, saves the plot as a JPEG file in the output directory (default: \code{TRUE}).}

\item{plot_width}{Sets the width (in inches) of the saved plot. Default: \code{11}}

\item{plot_height}{Sets the height (in inches) of the saved plot. Default: \code{8}.}

\item{plot_dpi}{Numeric. Sets the resolution (dots per inch) for the saved plot. Default: \code{600}.}

\item{...}{Additional arguments passed to other functions.}
}
\value{
A list containing:
\itemize{
\item \code{Diversity_Map}: A map with mean heterozygosity values for each location.
\item \code{Diversity_Table}: A data set containing the diversity data.
}
}
\description{
Calculate the heterozygosity for each SNP locus across samples in a data set, summarize
the genetic diversity (mean heterozygosity) for each location, and plot a proportional map of genetic
diversity based on the mean heterozygosity per location.
}
