% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiversityMap.R
\name{DiversityMap}
\alias{DiversityMap}
\title{DiversityMap}
\usage{
DiversityMap(df, SNP_Data, shapeFile, LongLat_data)
}
\arguments{
\item{df}{Final GRC dataframe}

\item{SNP_Data}{A dataframe of SNP data, where each row represents a sample and each column corresponds to a
genetic locus (e.g., "Pf3D7_"). The row names should correspond to the "Sample Internal ID".}

\item{LongLat_data}{Dataframe containing the locations and their geographical coo-ordinates, longitude and latitude}

\item{shapFile}{An \code{sf} object representing the shape file of the country}
}
\value{
A proportional map and a table of genetic diversity (mean heterozygosity) per location.
}
\description{
This function calculates the heterozygosity for each SNP locus across samples in a dataset, summarizes
the genetic diversity (mean heterozygosity) for each location, and plots a proportional map of genetic
diversity based on the mean heterozygosity per location.
}
\examples{
# Example usage:
DiversityMap(df = FinalData, SNP_Data = BarcodeData, shapeFile = GMB_shapefile, LongLat_data = Coordinates)
}
