% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Drug_Distribution_Plots.R
\name{drug_distribution_pm}
\alias{drug_distribution_pm}
\title{Drug Distribution Percentage Maps}
\usage{
drug_distribution_pm(
  df,
  drug_col,
  save_output = TRUE,
  period_name = "Full",
  map_data,
  time = NULL,
  label_size = 2.5,
  circle_num_size = 3.1,
  scale_circle_size = 10,
  ...
)
}
\arguments{
\item{df}{Final GRC data frame.}

\item{drug_col}{The name of the column representing the drug conditions (e.g., "Chloroquine" with
categories like Resistant, Mixed Resistant, and Sensitive).}

\item{save_output}{Logical. If \code{TRUE}, saves the plot as a JPEG file in the output directory (default: \code{FALSE}).}

\item{map_data}{A list containing the shape file and longitude-latitude data for mapping.}

\item{time}{Optional. A list defining time periods for filtering the data.}

\item{label_size}{Numeric. Controls the size of location labels on the map. Default: \code{2.5}.}

\item{circle_num_size}{Numeric. Controls the numbers inside the circles. Default: \code{3.1}}

\item{scale_circle_size}{Numeric. Scales the maximum circle size. Default: \code{10}.}
}
\value{
A list containing:
\itemize{
\item \code{Drug_Condition_Maps}: A list of ggplot objects for the drug condition maps.
\item \code{Drug_Condition_Table}: A data frame summarizing the drug resistance percentages and locations.
}
}
\description{
This function generates maps displaying the percentages of conditions
(All.Resistant, Mixed.Resistant, Resistant, Sensitive and All_Resistant) by location
for the selected drug column. Each circle houses the percentage of the condition for that particular location
and is labeled using the name and sample count for that location.
}
