% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter_Missigness_for_SNPs_Samples.R
\name{filter_snp_x_samples}
\alias{filter_snp_x_samples}
\title{Filter SNPs and Samples}
\usage{
filter_snp_x_samples(df, m_threshold)
}
\arguments{
\item{df}{Final GRC data frame}

\item{m_threshold}{A numeric value between 0 and 1, representing the maximum proportion of missing data
allowed for both SNPs and samples.}
}
\value{
The data frame with SNPs as column names and samples as row names,
filtered based on the specified missingness threshold.
}
\description{
This function filters out SNPs and samples based on a missingness threshold.
SNPs and samples with a proportion of missing data greater than the specified threshold will be removed.
}
