% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter_Missigness_for_SNPs_Samples.R
\name{Filter_SNPs_Samples}
\alias{Filter_SNPs_Samples}
\title{Filter_SNPs_Samples}
\usage{
Filter_SNPs_Samples(df, missingnessThreshold)
}
\arguments{
\item{df}{Final GRC dataframe}

\item{missingnessThreshold}{A numeric value between 0 and 1, representing the maximum proportion of missing data
allowed for both SNPs and samples. SNPs or samples with a higher proportion of missing
data will be filtered out.}
}
\value{
A dataframe with filtered SNPs and samples based on the specified missingness threshold.
}
\description{
This function filters out SNPs (variants) and samples from a genetic dataset based on a missingness threshold.
SNPs and samples with a proportion of missing data greater than the specified threshold will be removed.
}
\examples{
# Example usage:
Filter_SNPs_Samples(FinalData, 0.40)
}
