% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gene_Classifier.R
\name{Gene_Classifier}
\alias{Gene_Classifier}
\title{Classify Haplotypes for Drug Resistance}
\usage{
Gene_Classifier(df, drug_column = "Chloroquine")
}
\arguments{
\item{df}{A dataframe containing malaria sample data, including haplotype information.}

\item{drug_column}{A character string representing the drug resistance column to filter by (default is "Chloroquine").
Filtering will exclude samples marked as "Undetermined" or "Missing" in the selected column.}
}
\value{
A cleaned and classified dataframe with new columns for drug resistance status: Chloroquine, MultiDrug, Kelch13, Sulfadoxine, and Pyrimethamine.
}
\description{
This function classifies malaria sample data for drug resistance based on specific haplotypes.
It classifies resistance for Chloroquine, MultiDrug, Kelch13, Sulfadoxine, and Pyrimethamine,
allowing the user to filter based on a selected drug resistance column.
}
\examples{
clean_data <- Gene_Classifier(FinalData, drug_column = "Chloroquine", Country = "Gambia")

}
