% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate_IBS_Data.R
\name{generate_ibs_data}
\alias{generate_ibs_data}
\title{Generate IBS data frame}
\usage{
generate_ibs_data(df, snp_data, drug_col)
}
\arguments{
\item{df}{Final GRC data frame}

\item{snp_data}{A data frame of SNP data, where each row represents a sample and each column corresponds to a
genetic locus (e.g., "Pf3D7_"). The row names should correspond to the "Sample Internal ID".}

\item{drug_col}{The name of the column representing the drug conditions (e.g., "Chloroquine" with
categories like Resistant, Mixed Resistant, and Sensitive).}
}
\value{
A list containing a histogram of IBS scores and data frame containing the IBS scores between sample pairs along with associated metadata.
The data frame will include the following columns:
\itemize{
\item \code{S1}: First sample in the pair
\item \code{S2}: Second sample in the pair
\item \code{value}: IBS score between S1 and S2
\item \code{LS1}: Location of sample S1
\item \code{LS2}: Location of sample S2
\item \code{DCS1}: Drug condition for sample S1 (e.g., Chloroquine)
\item \code{DCS2}: Drug condition for sample S2
\item \code{RgS1}: Region for the location of sample S1
\item \code{RgS2}: Region for the location of sample S2

}
}
\description{
This function calculates identity-by-state (IBS) scores between samples based on the barcode_data sequence which is made
up of the SNPs columns in th GRC Data. It reshapes the IBS matrix into a long format and annotates the data with metadata,
including sample locations, drug conditions, location of the samples pairing, etc.
}
\examples{
# Example usage:
generate_ibs_data(df = GRC_Data, snp_data = BarcodeData, drug_col = "Chloroquine")
head(IBS_Data)
}
