% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{generate_location_colors}
\alias{generate_location_colors}
\title{Generate a Color Palette for Locations}
\usage{
generate_location_colors(df, location_col = "Location")
}
\arguments{
\item{df}{A data frame containing a \code{Location} column.}

\item{location_col}{The name of the column representing locations (default is "Location").}
}
\value{
A named vector of colors where names correspond to unique locations.
}
\description{
This function assigns colors to unique locations in a GRC data frame
}
