% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate_IBS_Data.R
\name{GenerateIBS_Data}
\alias{GenerateIBS_Data}
\title{GenerateIBS_Data}
\usage{
GenerateIBS_Data(df, SNP_Data, LongLat_data)
}
\arguments{
\item{df}{Final GRC dataframe}

\item{SNP_Data}{A dataframe of SNP data, where each row represents a sample and each column corresponds to a
genetic locus (e.g., "Pf3D7_"). The row names should correspond to the "Sample Internal ID".}

\item{LongLat_data}{A dataframe containing the longitude, latitude, and region information for each location.
It should have columns for "Location" and "Regions".}
}
\value{
A list containing a histogram of IBS scores and dataframe in long format containing the IBS scores between sample pairs along with associated metadata.
The dataframe will include the following columns:
\itemize{
\item \code{S1}: First sample in the pair
\item \code{S2}: Second sample in the pair
\item \code{value}: IBS score between S1 and S2
\item \code{LS1}: Location of sample S1
\item \code{LS2}: Location of sample S2
\item \code{DCS1}: Drug condition for sample S1 (e.g., Chloroquine)
\item \code{DCS2}: Drug condition for sample S2
\item \code{RgS1}: Region for the location of sample S1
\item \code{RgS2}: Region for the location of sample S2

}
}
\description{
This function calculates identity-by-state (IBS) scores between samples based on SNP data. It reshapes the
IBS matrix into a long format and annotates the data with metadata, including sample locations, drug
conditions, and regions.
}
\examples{
# Example usage:
GenerateIBS_Data(df = FinalData, SNP_Data = BarcodeData, LongLat_data = Coordinates)
head(IBS_Data)
}
