% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageLibraries.R
\docType{package}
\name{grcMPB-package}
\alias{grcMPB}
\alias{grcMPB-package}
\title{grcMPB: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Brandon Ngwa \email{ngwabrandon10@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
