% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Haplotype_Frequency.R
\name{haplotype_frequency}
\alias{haplotype_frequency}
\title{Haplotype Frequency Plots}
\usage{
haplotype_frequency(
  df,
  gene_col,
  filter_gene_col = FALSE,
  save_output = TRUE,
  label_repel = 1.3,
  period_name = "Full",
  label_size = 2.5,
  map_data,
  donut_chart_size = 0.5,
  time = NULL,
  plot_width = 11,
  plot_height = 6,
  plot_dpi = 600,
  ...
)
}
\arguments{
\item{df}{Combined GRC data frame}

\item{gene_col}{The column containing haplotype data (e.g., "PfCRT")}

\item{filter_gene_col}{A Boolean (TRUE or FALSE). If TRUE, samples with missen haplotypes are filtered out of the data set.}

\item{save_output}{Logical. If \code{TRUE}, saves the plot as a JPEG file in the output directory (default: \code{FALSE}).}

\item{label_repel}{Numeric. Controls the distance of the label from the points on the map. Default: \code{1.3}.}

\item{period_name}{The period name for the plot. Defualt: \code{FULL}}

\item{label_size}{Numeric. Controls the size of location labels on the map. Default: \code{2.5}.}

\item{map_data}{A list containing the shape file and longitude-latitude data for mapping.}

\item{donut_chart_size}{Numeric. Scales the maximum pie chart size. Default: \code{0.5}}

\item{time}{Optional. A list defining time periods for filtering the data.}

\item{plot_width}{Sets the width (in inches) of the saved plot. Default: \code{11}.}

\item{plot_height}{Sets the height (in inches) of the saved plot. Default: \code{6}.}

\item{plot_dpi}{Numeric. Sets the resolution (dots per inch) for the saved plot. Default: \code{600}.}

\item{...}{Additional arguments passed to other functions.}
}
\value{
A list containing:
\itemize{
\item \code{Bar_Chart}: A bar chart of haplotype frequencies with percentages.
\item \code{Donut_Chart}: A pie chart map of haplotype frequency by location.
\item \code{Haplotype_Summary_Table}: A data frame summarizing the counts and percentages of haplotypes,
including an "Others" category for low-frequency haplotypes.
}
}
\description{
This function generates the frequency of unique haplotypes for a specific gene (e.g, PfCRT)
}
