% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Haplotype_Proportion_Plots.R
\name{Haplotype_Proportion}
\alias{Haplotype_Proportion}
\title{Haplotype Proportion Plots}
\usage{
Haplotype_Proportion(
  df,
  gene_col,
  drug_col,
  saveOutput = TRUE,
  period_name = "Full",
  labelSize = 2.5,
  scaleCircleSize = 0.035,
  time = NULL,
  ...
)
}
\arguments{
\item{df}{Final GRC dataframe}

\item{gene_col}{The column containing haplotype data (e.g., "PfCRT")}

\item{drug_col}{The name of the column representing the drug conditions}

\item{saveOutput}{Logical. Whether to save the output plots to files (default is TRUE)}

\item{labelSize}{Size of the labels on the map}

\item{scaleCircleSize}{Used to scale the pie chart circle sizes}

\item{time}{Optional. A list defining time periods, where each list element contains:}
}
\description{
This function generates haplotype proportion visualizations (bar chart and pie chart map)
with flexible time-based filtering and output saving capabilities.
}
\examples{
Haplotype_Proportion(GRC_data, drug_col = "Chloroquine", gene_col = "PfCRT")

}
