% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectednessMap.R
\name{IBS_ConnectednesMaps}
\alias{IBS_ConnectednesMaps}
\title{Plot IBS Connectedness Maps}
\usage{
IBS_ConnectednesMaps(
  shapeFile,
  LongLatData,
  LongLatData_sf,
  IbsData,
  prop_column,
  saveName,
  plotName,
  limits = c(0, max(IbsData[[prop_column]])),
  breaks = seq(0, max(IbsData[[prop_column]]), 1),
  location_col = "Location",
  long_col = "long",
  lat_col = "lat"
)
}
\arguments{
\item{shapeFile}{sf object representing the base map for plotting geographic areas.}

\item{LongLatData_sf}{The LongLat_Data converted to a sf object.}

\item{IbsData}{Data frame containing the pairwise IBS data between locations, along with the computed percentages.}

\item{prop_column}{Character string specifying the name of the column in \code{IbsData} that contains the pairwise percentages (e.g., "0.75_PairCountGroup.per").}

\item{saveName}{Character string for naming the output plot file.}

\item{plotName}{Character string for the title of the plot.}

\item{limits}{Numeric vector of length 2 specifying the range of values for the color scale and line width (default is \code{c(0, max(dataframe[[prop_column]]))}).}

\item{breaks}{Numeric vector indicating the break points for the color scale and line width legend (default is a sequence from 0 to the maximum of \code{prop_column}).}

\item{location_col}{Character string for the column name in \code{LongLatData} representing the locations (default is "Location").}

\item{long_col}{Character string for the column name in \code{LongLatData} representing the longitude (default is "long").}

\item{lat_col}{Character string for the column name in \code{LongLatData} representing the latitude (default is "lat").}

\item{LongLat_Data}{Dataframe containing the locations and their geographical coo-ordinates, longitude and latitude}
}
\value{
A ggplot2 plot object showing the connectedness map.
}
\description{
This function generates connectedness maps showing connections between geographic locations based on the pairwise IBS percentages calculated between them. The connections are visualized as curved lines, where the width and color of the lines reflect the strength of the pair percentage.
}
\details{
The function creates a map using geographic shapes (\code{shapeFile}) and overlays lines (\code{geom_curve}) to represent the connection between two geographic locations. The line's thickness and color correspond to the pairwise percentage columns .

The function saves the plot as a JPEG file in a predefined directory (\code{IBS_ConnectednessMaps}).
}
\examples{
# Example usage:
IBS_ConnectednesMaps(shapeFile = shapeFile, LongLatData_sf = LongLatData_sf,
                     LongLatData = LongLatData, IbsData = IbsData,
                     prop_column = "0.75_PairCountGroup.per",
                     saveName = "IBS_Map", plotName = "IBS Connectedness")

}
