% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summarise_IBS_Data.R
\name{IBS_DataSummary}
\alias{IBS_DataSummary}
\title{Manipulate the IBS Group Dataframes}
\usage{
IBS_DataSummary(MeltIbsData, LongLat_data, IBS_Threshold = 0.75)
}
\arguments{
\item{MeltIbsData}{A dataframe containing the melted IBS data. This dataframe should include the columns \code{LS1}, \code{LS2}, and \code{value}, where \code{value} represents the IBS score between two locations.}

\item{LongLat_data}{A dataframe containing the latitude and longitude for each location. The dataframe should include the columns \code{Location}, \code{lat}, and \code{long}.}

\item{IBS_Threshold}{A numeric value between 0 and 1 representing the IBS threshold for pair counts. Default is 0.75. The function dynamically names columns based on this threshold.}
}
\value{
A dataframe summarizing the IBS scores between location pairs. The output dataframe contains the following columns:
\itemize{
\item \code{LS1, LS2}: The locations in each pair.
\item \code{mean_IBS}: The mean IBS score for the pair.
\item \code{median_IBS}: The median IBS score for the pair.
\item \code{TotalPairCount}: The total number of pairs between the two locations.
\item \code{<IBS_Threshold>_PairCount}: The number of pairs with an IBS score greater than or equal to the threshold.
\item \code{TotalPairCount_per}: The percentage of total pairs in the dataset.
\item \code{<IBS_Threshold>_PairCountGroup.per}: The percentage of total pairs that meet or exceed the IBS threshold.
\item \code{<IBS_Threshold>_PairCountLocation.per}: The percentage of pairs for each location that meet or exceed the IBS threshold.
\item \code{<IBS_Threshold>_PairCountAll.per}: The percentage of total pairs that meet or exceed the IBS threshold across all locations.
\item \code{Lat1, Long1, Lat2, Long2}: The latitude and longitude for both locations in each pair.
}
}
\description{
This function processes and summarizes a melted IBS dataframe. It calculates the mean and median IBS values for each location pair, counts the total number of pairs, and determines the number of pairs above a user-defined IBS threshold. The function also calculates various percentages and appends geographical coordinates for each location.
}
\examples{
# Example usage:
# Assuming MeltIbsData and LongLat_data are already defined:
IBS_DataSummary(MeltIbsData, LongLat_data, IBS_Threshold = 0.75)
}
