% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{initialize_output_paths}
\alias{initialize_output_paths}
\title{Initialize Output Paths}
\usage{
initialize_output_paths(dir1 = NULL, dir2 = NULL)
}
\arguments{
\item{dir1}{(Optional) Name for the first sub directory within Outputs.}

\item{dir2}{(Optional) Name for the nested sub directory within dir1.}
}
\value{
The path to the directory where outputs can be saved.
}
\description{
A helper function to create output directories and nested directories based on the user's specifications.
This function is internal and should not be called directly by the user.
}
