% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{mapping_data}
\alias{mapping_data}
\title{Mapping data helper function}
\usage{
mapping_data(df, shapefile, long_lat_data, location_col, long_col, lat_col)
}
\arguments{
\item{df}{Combined GRC data frame}

\item{shapefile}{An \code{sf} object representing the shape file of the country.}

\item{long_lat_data}{A dataframe containing the locations and their geographical coordinates (longitude and latitude).}

\item{location_col}{The name of the column representing location identifiers.}

\item{long_col}{The name of the longitude column.}

\item{lat_col}{The name of the latitude column.}
}
\value{
A list containing the \code{shapefile}, standardized \code{long_lat_data} with an added color column,
and \code{location_colors} mapping for use in other visualizations.
}
\description{
Standardizes the structure of location data for mapping functions by checking and renaming specified columns.
The function also assigns colors to each location which can be used in subsequent plots.
}
