% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mutation_Frequency_Plots.R
\name{mutation_frequency}
\alias{mutation_frequency}
\title{Mutation Frequency Calculation and Visualization}
\usage{
mutation_frequency(
  df,
  gene_col,
  filter_gene_col = FALSE,
  save_output = TRUE,
  period_name = "Full",
  time = NULL,
  map_data,
  label_repel = 1.3,
  label_size = 2.5,
  circle_num_size = 3.1,
  scale_circle_size = 10,
  include_mixed = FALSE,
  plot_width = 11,
  plot_height = 6,
  plot_dpi = 600,
  ...
)
}
\arguments{
\item{df}{Combined GRC data frame}

\item{gene_col}{The column containing haplotype data (e.g., "PfCRT", "PfDHPS", "PfDHFR", "PfMDR1")}

\item{filter_gene_col}{A Boolean (TRUE or FALSE). If TRUE, samples with missen haplotypes are filtered out of the data set.}

\item{save_output}{Logical. If \code{TRUE}, saves the plot as a JPEG file in the output directory (default: \code{FALSE}).}

\item{period_name}{The period name for the plot. Defualt: \code{FULL}}

\item{time}{Optional. A list defining time periods.}

\item{map_data}{A list containing the shape file and longitude-latitude data for mapping.}

\item{label_repel}{Numeric. Controls the distance of the label from the points on the map. Default: \code{1.3}.}

\item{label_size}{Numeric. Controls the size of location labels on the map. Default: \code{2.5}.}

\item{circle_num_size}{Numeric. Controls the numbers inside the circles. Default: \code{3.1}.}

\item{scale_circle_size}{Numeric. Scales the maximum circle size. Default: \code{10}.}

\item{include_mixed}{Boolean, whether to include mixed haplotypes in counts}

\item{plot_width}{Sets the width (in inches) of the saved plot. Default: \code{11}.}

\item{plot_height}{Sets the height (in inches) of the saved plot. Default: \code{6}.}

\item{plot_dpi}{Numeric. Sets the resolution (dots per inch) for the saved plot. Default: 600.}

\item{...}{Additional arguments passed to other functions.}
}
\description{
This function calculates the frequency of wild, mutated, mixed, and missing haplotypes for a specified gene.
It also generates two tables: one for mutation types by position and another for mutations by location.
The function provides the option to include mixed haplotypes in the counts for both tables.
The first plot is a bar chart with the percentages of, wild, mutated, mixed and missing for the haplotype.
The second plots are maps which show these percentages in circles for all available locations.
}
