% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_Plots.R
\name{network_plot}
\alias{network_plot}
\title{Generate Network Plots using the IBS matrix}
\usage{
network_plot(
  ibs_matrix,
  df,
  ibs_th = 0.75,
  nodepoint_size = 1.5,
  edge_arc_strength = 0.4,
  save_output = FALSE,
  drug_col = NULL,
  plot_width = 8,
  plot_height = 3,
  plot_dpi = 600,
  location_colors
)
}
\arguments{
\item{ibs_matrix}{Matrix of IBS scores for all sample pairs. Row and column names should match \verb{Sample Internal ID}.}

\item{df}{Combined GRC data frame}

\item{ibs_th}{Minimum Ibs threshold to show connections. All other pairs with less that this threshold will show no connection.}

\item{nodepoint_size}{Numeric. Controls the size of the node points on the plot.}

\item{edge_arc_strength}{Numeric. Controls the curviture of the edges or lines conneting the nodes.}

\item{save_output}{Logical. If \code{TRUE}, saves the plot as a JPEG file in the output directory (default: \code{FALSE}).}

\item{drug_col}{The name of the column representing the drug conditions (e.g., "Chloroquine" with
categories like Resistant, Mixed Resistant, and Sensitive).}

\item{plot_width}{Sets the width (in inches) of the saved plot. Default: \code{25}.}

\item{plot_height}{Sets the height (in inches) of the saved plot. Default: \code{18}.}

\item{plot_dpi}{Numeric. Sets the resolution (dots per inch) for the saved plot. Default: \code{600}.}

\item{location_colors}{Named vector of colors for locations, typically from mapping_data$location_colors.}
}
\value{
A list containing the network plots.
}
\description{
The \code{network_plot} function creates a neighbor joining tree using the Ibs matrix.
}
\details{
The function builds a network plot joining tree using the Ibs matrix.
If \code{drug_col} is specified, the tree is generated for each category in the drug column.
}
