% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NJ_Trees.R
\name{nj_tree}
\alias{nj_tree}
\title{Generate Neighbor Joining Trees using the IBS matrix}
\usage{
nj_tree(
  ibs_matrix,
  df,
  tippoint_size = 4,
  line_size = 0.6,
  save_output = FALSE,
  drug_col = NULL
)
}
\arguments{
\item{ibs_matrix}{Matrix of IBS scores for all sample pairs. Row and column names should match \verb{Sample Internal ID}.}

\item{df}{Final GRC data frame}

\item{tippoint_size}{Numeric. Controls the size of the tip points on the tree.}

\item{line_size}{Numeric. Controls the size of the branches or lines.}

\item{save_output}{Logical. If \code{TRUE}, saves the plot as a JPEG file in the output directory (default: \code{FALSE}).}

\item{drug_col}{The name of the column representing the drug conditions (e.g., "Chloroquine" with
categories like Resistant, Mixed Resistant, and Sensitive).}
}
\value{
A list containing the neighbor joining trees.
}
\description{
The \code{nj_tree} function creates a neighbor joining tree using the Ibs matrix.
}
\details{
The function builds a neighbor joining tree using the Ibs matrix after converting it to a distance matrix.
It then using the the function nj() from the ape package to build the tree. After this we use package ggtree
from the BioManager package to visualize our tree.
If \code{drug_col} is specified, the tree is generated for each category in th drug column.
}
\examples{
nj_tree(ibs_matrix = Ibs_Matrix,
        df = GRC Data,
        drug_col = "Chloroquine")


}
