% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Drug_Condition_Proportion_Map.R
\name{Proportion_Map}
\alias{Proportion_Map}
\title{Drug Condition Map}
\usage{
Proportion_Map(
  shapeFile,
  summaryData,
  summaryData_sf,
  prop_column,
  saveLocation = NULL,
  location_col = "Location",
  long_col = "long",
  lat_col = "lat"
)
}
\arguments{
\item{shapeFile}{An \code{sf} object representing the shape file of the country}

\item{summaryData}{This is the summaryTable data which contains the information on the drug conditions}

\item{summaryData_sf}{This is the summaryTable data turned into a sf object so we can use it on the map}

\item{prop_column}{This is the drug condition percentage column your want to use to plot on the map ("Mixed.Resistant.per","Resistant.per","Sensitive.per","All_Resistant.per")}

\item{saveLocation}{Loacation path to save your plot}

\item{location_col}{The column containing the locations in your dataframe}

\item{long_col}{The column containing the longitude information for you location}

\item{lat_col}{The column containing the latitude information for your location}
}
\value{
The function saves the drug condition map for the selected prop_column
}
\description{
This function generates a map displaying the total sample counts for each location_col based on long_colitude and lat_colitude coordinates.
It creates a proportional symbol map with sample counts represented by varying symbol sizes, and it labels each location_col with the total sample count.
The map is saved as a JPEG file in the \code{savePath} directory.
}
