% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleCountMap.R
\name{SampleCountMap}
\alias{SampleCountMap}
\title{Sample Count Map}
\usage{
SampleCountMap(
  df,
  drug_col,
  breaks = NULL,
  saveOutput = TRUE,
  period_name = "Full",
  labelSize = 2.5,
  scaleCircleSize = 11,
  time = NULL,
  ...
)
}
\arguments{
\item{df}{Final GRC dataframe.}

\item{drug_col}{The name of the column representing the drug conditions (e.g., "Chloroquine").}

\item{breaks}{Used to set the breaks for the circle sizes}

\item{saveOutput}{Logical. Whether to save the output plots to files (default is FALSE).}

\item{labelSize}{Used to set the size of th labels on the map}

\item{scaleCircleSize}{Used to scale the circle sizes.}

\item{time}{Optional. A list defining time periods, where each list element contains:
\itemize{
\item \code{type}: Either "year" or "period" to define time scope.
\item \code{start}: The start year of the time period.
\item \code{end}: The end year of the time period (for periods only).
\item \code{name}: The label to use for the period.
}

Example for generating time-period-based summary
time_periods <- list(list(type = "year", start = 2010, name = "2010"),
list(type = "period", start = 2015, end = 2019, name = "2015-2019"))}
}
\description{
This function generates a map displaying the Sample_Count sample counts for each location based on longitude and latitude coordinates.
It creates a proportional symbol map with sample counts represented by varying symbol sizes, and labels each location with the Sample_Count sample count.
The map is saved as a JPEG file in the \code{savePath} directory if \code{saveOutput} is set to \code{TRUE}.
}
\examples{
# Ensure that df, shapeFile, and LongLat_data are properly prepared in your global environment
SampleCountMap(GRC_data, drug_col = "Chlroquine",)

}
