% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{split_haplotype}
\alias{split_haplotype}
\title{Split Haplotype}
\usage{
split_haplotype(haplotype)
}
\arguments{
\item{haplotype}{haplotype to string to split  (e.g., "CVIE\[M/T\]" )}
}
\value{
a vector with of each element in the splitted string.
}
\description{
A helper function to split the haplotypes combination into a vector if strings
}
