% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{TemporalData_List}
\alias{TemporalData_List}
\title{Process Data for Temporal Analysis}
\usage{
TemporalData_List(df, func, time, ...)
}
\arguments{
\item{df}{The input dataframe containing the temporal data}

\item{func}{The function to apply to each temporal subset}

\item{time}{List of time periods, where each element contains:
\itemize{
\item type: Either "year" or "period"
\item start: Start year
\item end: End year (for periods only)
\item name: Label for the period
}}

\item{...}{Additional arguments passed to func}

\item{location_col}{Name of the location column}

\item{drug_col}{Name of the drug column}

\item{LongLat_data}{Dataframe containing geographical coordinates}
}
\value{
A list containing results for each time period
}
\description{
A helper function that processes data based on specified time periods and applies
a given function to each temporal subset. This function standardizes temporal analysis across
different visualization functions.
}
