% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Combine_GRC_Sheets.R
\name{combine_grc_sheets}
\alias{combine_grc_sheets}
\title{Combine GRC Sheets}
\usage{
combine_grc_sheets(
  input_folder,
  filter_pf = TRUE,
  country = NULL,
  save_output = TRUE,
  output_dir = NULL
)
}
\arguments{
\item{input_folder}{Path to the folder containing the GRC Excel files.}

\item{filter_pf}{Logical. If \code{TRUE}the data is filtered for only \verb{pf(plasmoduim falciparum)}specie.}

\item{country}{A character string representing the country for which specific data cleaning rules should be applied (optional).
If "Gambia", location names will be re coded and some locations will be filtered out.}

\item{save_output}{Logical. Whether to save the output plots to the Outputs folder (default is FALSE).}

\item{output_dir}{Output directory to store your results.}
}
\value{
A data frame containing multiple GRC excel sheets merged together.
}
\description{
This function merges data from multiple GRC Excel files, cleans the data and filters by country if specified.
If save_output equals to TRUE then a folder called Outputs in created in the working directory which will host all the outputs.
Outputs is also passed to your working environment as a list containing the path to your output folder.
}
