% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Combine_GRC_Sheets.R
\name{Combine_GRC_Sheets}
\alias{Combine_GRC_Sheets}
\title{Combine GRC Sheets Function}
\usage{
Combine_GRC_Sheets(input_folder, Country = NULL, saveOuput = TRUE)
}
\arguments{
\item{input_folder}{Path to the folder containing the GRC Excel files.}

\item{saveOuput}{This allows th users to save their ouputs or not, taking in the Values TRUE or False with TRUE as the default}

\item{country}{A character string representing the country for which specific data cleaning rules should be applied (optional).
If "Gambia", location names will be re coded and some locations will be filtered out.}
}
\value{
A data frame containing the merged data.
}
\description{
This function merges data from multiple GRC Excel files, cleans the data and filters by country if specified.
}
