% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Haplotype_Proportion_Plots.R
\name{create_haplotype_plots}
\alias{create_haplotype_plots}
\title{Internal function to create the haplotype plots}
\usage{
create_haplotype_plots(
  df,
  gene_col,
  drug_col,
  save_output = TRUE,
  period_name = "Full",
  label_size = 2.5,
  map_data,
  sacle_piechart_size = 0.035,
  ...
)
}
\arguments{
\item{df}{Final GRC data frame}

\item{gene_col}{The column containing haplotype data (e.g., "PfCRT")}

\item{drug_col}{The name of the column representing the drug conditions (e.g., "Chloroquine" with
categories like Resistant, Mixed Resistant, and Sensitive).}

\item{save_output}{Logical. If \code{TRUE}, saves the plot as a JPEG file in the output directory (default: \code{FALSE}).}

\item{label_size}{Numeric. Controls the size of location labels on the map. Default: \code{2.5}.}

\item{map_data}{A list containing the shape file and longitude-latitude data for mapping.}

\item{sacle_piechart_size}{Numeric. Scales the maximum pie chart size. Default: \code{0.035}}
}
\description{
Internal function to create the haplotype plots
}
\keyword{internal}
