% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleCountMap.R
\name{create_sc_map}
\alias{create_sc_map}
\title{Internal Function to Create Sample Count Map}
\usage{
create_sc_map(
  df,
  map_data,
  breaks = NULL,
  save_output = TRUE,
  period_name = "Full",
  label_size = 2.5,
  scale_circle_size = 11,
  ...
)
}
\arguments{
\item{df}{Final GRC data frame.}

\item{map_data}{A list containing the shape file and longitude-latitude data for mapping.}

\item{breaks}{Numeric vector. Sets the breaks for circle sizes. Default: \code{c(10, 50, 100, 200, 300, 400, 500)}.}

\item{save_output}{Logical. If \code{TRUE}, saves the plot as a JPEG file in the output directory (default: \code{FALSE}).}

\item{label_size}{Numeric. Controls the size of location labels on the map. Default: \code{2.5}.}

\item{scale_circle_size}{Numeric. Scales the maximum circle size. Default: \code{11}.}
}
\description{
Internal Function to Create Sample Count Map
}
\keyword{internal}
