% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiversityMap.R
\name{diversity_map}
\alias{diversity_map}
\title{Create Diversity Map}
\usage{
diversity_map(
  df,
  snp_data,
  period_name = "Full",
  label_size = 2.5,
  time = NULL,
  map_data,
  circle_num_size = 3.1,
  scale_circle_size = 10,
  save_output = TRUE,
  ...
)
}
\arguments{
\item{df}{Final GRC data frame}

\item{snp_data}{A data frame of SNP data, where each row represents a sample and each column corresponds to a
genetic locus (e.g., "Pf3D7_"). The row names should correspond to the "Sample Internal ID".}

\item{label_size}{Numeric. Controls the size of location labels on the map. Default: \code{2.5}.}

\item{time}{Optional. A list defining time periods for filtering the data.}

\item{map_data}{A list containing the shape file and longitude-latitude data for mapping.}

\item{circle_num_size}{Numeric. Controls the numbers inside the circles. Default: \code{3.1}.}

\item{scale_circle_size}{Numeric. Scales the maximum circle size. Default: \code{10}.}

\item{save_output}{Logical. If \code{TRUE}, saves the plot as a JPEG file in the output directory (default: \code{FALSE}).}

\item{drug_col}{The name of the column representing the drug conditions (e.g., "Chloroquine" with
categories like Resistant, Mixed Resistant, and Sensitive).}
}
\value{
A proportional map and a table of genetic diversity (mean heterozygosity) per location.

A list containing:
\itemize{
\item \code{Diversity_Map}: A map with mean heterozygosity values for each location.
\item \code{Diversity_Table}: A data set containing the diversity data.
}
\description{
This function calculates the heterozygosity for each SNP locus across samples in a dataset, summarizes
the genetic diversity (mean heterozygosity) for each location, and plots a proportional map of genetic
diversity based on the mean heterozygosity per location.
}
\examples{
diversity_map(df = GRC_Data,
drug_col = "Chloroquine",
snp_data = barcode_data)

}
