% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiversityMap.R
\name{DiversityMap}
\alias{DiversityMap}
\title{DiversityMap}
\usage{
DiversityMap(
  df,
  drug_col,
  SNP_Data,
  period_name = "Full",
  label_size = 2.5,
  time = NULL,
  circle_num_size = 3.1,
  scale_circle_size = 10,
  saveOutput = TRUE,
  ...
)
}
\arguments{
\item{df}{Final GRC dataframe}

\item{drug_col}{The name of the column representing the drug conditions (e.g., "Chloroquine").}

\item{SNP_Data}{A dataframe of SNP data, where each row represents a sample and each column corresponds to a
genetic locus (e.g., "Pf3D7_"). The row names should correspond to the "Sample Internal ID".}

\item{label_size}{Used to set the size of the labels on the map.}

\item{time}{Optional. A list defining time periods.}

\item{circle_num_size}{Used to set th sizes of the numbers in th circles.}

\item{scale_circle_size}{Used to scale the size of th circles.}

\item{saveOutput}{Logical. Whether to save the output plots to files (default is FALSE).}
}
\value{
A proportional map and a table of genetic diversity (mean heterozygosity) per location.
}
\description{
This function calculates the heterozygosity for each SNP locus across samples in a dataset, summarizes
the genetic diversity (mean heterozygosity) for each location, and plots a proportional map of genetic
diversity based on the mean heterozygosity per location.
}
\examples{
# Example usage:
DiversityMap(df = FinalData, drug_col = "Chloroquine", SNP_Data = BarcodeData)

}
