% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Drug_Condition_Proportion_Charts.R
\name{Drug_Distribution}
\alias{Drug_Distribution}
\title{Generate and Visualize Drug Distribution Data}
\usage{
Drug_Distribution(
  df,
  drug_col,
  save_output = TRUE,
  time = NULL,
  period_name = "Full",
  colors = c(Resistant = "#525CEB", Mixed.Resistant = "#808000", Sensitive = "#800000"),
  ...
)
}
\arguments{
\item{df}{Final GRC data frame.}

\item{drug_col}{The name of the column representing the drug conditions (e.g., "Chloroquine" with conditions, Resistant, mixed resistant and sensitive).}

\item{save_output}{Logical. Whether to save the output plots to files (default is FALSE).}

\item{time}{Optional. A list defining time periods, where each list element contains:}

\item{colors}{A three color pallet for th drug_col conditions.}
}
\description{
This function generates bar charts and summary data of drug conditions for s specific drug,(e.g, Chloroquine)
It provides returns two bar charts one for the the overall proportion of the drug condition for th entire data set
and another for each location.
}
