% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Drug_Distribution_Plots.R
\name{drug_distribution}
\alias{drug_distribution}
\title{Drug Distribution Plots}
\usage{
drug_distribution(
  df,
  drug_col,
  save_output = TRUE,
  time = NULL,
  period_name = "Full",
  ...
)
}
\arguments{
\item{df}{Final GRC data frame.}

\item{drug_col}{The name of the column representing the drug conditions (e.g., "Chloroquine" with
categories like Resistant, Mixed Resistant, and Sensitive).}

\item{save_output}{Logical. If \code{TRUE}, saves the plot as a JPEG file in the output directory (default: \code{FALSE}).}

\item{time}{Optional. A list defining time periods for filtering the data.}

\item{period_name}{The period name for the plot. Defualt: \code{FULL}}

\item{...}{Additional arguments passed to other functions.}
}
\value{
A list containing:
\itemize{
\item \code{Bar1}: A bar plot showing the distribution of conditions across all locations.
\item \code{Bar2}: A bar plot showing the distribution of conditions grouped by location.
\item \code{Condition_Table1}: A summary table showing the count of each condition for each location.
\item \code{Condition_Table2}: A summary table showing the percentage of each condition for the entire data set.
}
}
\description{
This function generates bar charts and summary data of drug conditions for a specific drug,(e.g, Chloroquine)
in the data set and another showing the distribution for each location.
}
