% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Drug_Condition_Proportion_Charts.R
\name{Drug_Distribution}
\alias{Drug_Distribution}
\title{Generate Drug Status Bar Charts and Save Summary Data}
\usage{
Drug_Distribution(df, LongLat_data, drug_col, time = NULL, saveOuput = TRUE)
}
\arguments{
\item{df}{Final GRC dataframe}

\item{LongLat_data}{Dataframe containing the locations and their geographical coordinates, longitude and latitude}

\item{drug_col}{The name of the column representing the drug conditions (e.g., "Chloroquine").}

\item{saveOuput}{This allows th users to save their ouputs or not, taking in the Values TRUE or False with TRUE as the default}
}
\value{
A list containing two ggplot objects: \code{bar1} and \code{bar2}.
}
\description{
This function generates bar charts for drug status based on a specified drug column
from the provided data and saves the summary data to the global environment.
}
