% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gene_Classifier.R
\name{gene_classifier}
\alias{gene_classifier}
\title{Classify Haplotypes for Drug Resistance}
\usage{
gene_classifier(df, drug_column = "Chloroquine", save_output = FALSE)
}
\arguments{
\item{df}{Final GRC data frame}

\item{drug_column}{A character string representing the drug column to filter by (default is "Chloroquine").
Filtering will exclude samples marked as "Undetermined" or "Missing" in the selected column.}

\item{save_output}{Logical. Whether to save the outputs to the Outputs folder (default is FALSE).}
}
\value{
A GRC data frame with new columns for drugs:
Chloroquine, MultiDrug, Kelch13, Sulfadoxine, and Pyrimethamine.
}
\description{
This function uses the haplotypes in the gene columns, (PfCRT, PfMDR1, Kelch, PfDHPS and PfDHFR ) and classifies them as
Resistant, Mixed Resistant, Sensitive, missing and undetermined for their specific associated drugs.
It creates new columns containing this categories and names them with the drugs they are associate with.
E.g
\itemize{
\item \code{Chloroquine}: for PfCRT.
\item \code{MultiDrug}: for PfMDR1.
\item \code{Kelch13}: for Kelch.
\item \code{Sulfadoxine}: for PfDHPS.
\item \code{Pyrimethamine}: for PfDHFR.
}
After this it then filters out missing and undetermined cases for a specific drug determined on which the user wants to work with.
}
\examples{
gene_classifier(df = GRC_Data, drug_column = "Chloroquine")

}
