% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Haplotype_Proportion_Plots.R
\name{Haplotype_Proportion}
\alias{Haplotype_Proportion}
\title{Haplotype Proportion Function}
\usage{
Haplotype_Proportion(df, shapeFile, LongLat_data, col_gene)
}
\arguments{
\item{df}{Final GRC dataframe}

\item{LongLat_data}{Dataframe containing the locations and their geographical coo-ordinates, longitude and latitude}

\item{col_gene}{The column in the df that contains haplotype data (e.g., "PfCRT").}

\item{shapFile}{An \code{sf} object representing the shape file of the country}
}
\value{
A bar chart of haplotype proportions and a pie chart map saved as JPEGs.
}
\description{
This function generates a bar chart of haplotype proportions and a pie chart map
representing haplotype proportions across locations. It also saves the generated plots to a specified path.
}
\examples{
Haplotype_Proportion(FinalData, GMB, LongLat, "PfCRT")

}
