% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Haplotype_Proportion_Plots.R
\name{Haplotype_Proportion}
\alias{Haplotype_Proportion}
\title{Haplotype Proportion Plots}
\usage{
Haplotype_Proportion(
  df,
  gene_col,
  drug_col,
  save_output = TRUE,
  period_name = "Full",
  label_size = 2.5,
  scale_circle_size = 0.035,
  time = NULL,
  ...
)
}
\arguments{
\item{df}{Final GRC dataframe}

\item{gene_col}{The column containing haplotype data (e.g., "PfCRT")}

\item{drug_col}{The name of the column representing the drug conditions}

\item{save_output}{Logical. Whether to save the output plots to files (default is TRUE)}

\item{label_size}{Size of the labels on the map}

\item{scale_circle_size}{Used to scale the pie chart circle sizes}

\item{time}{Optional. A list defining time periods, where each list element contains:}
}
\description{
This function generates the proportion of unique haplotypes for a specific gene (e.g, PfCRT)
}
\examples{
Haplotype_Proportion(GRC_data, drug_col = "Chloroquine", gene_col = "PfCRT")

}
