% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Haplotype_Proportion_Plots.R
\name{haplotype_proportion}
\alias{haplotype_proportion}
\title{Haplotype Proportion Plots}
\usage{
haplotype_proportion(
  df,
  gene_col,
  drug_col,
  save_output = TRUE,
  period_name = "Full",
  label_size = 2.5,
  map_data,
  sacle_piechart_size = 0.035,
  time = NULL,
  ...
)
}
\arguments{
\item{df}{Final GRC data frame}

\item{gene_col}{The column containing haplotype data (e.g., "PfCRT")}

\item{drug_col}{The name of the column representing the drug conditions (e.g., "Chloroquine" with
categories like Resistant, Mixed Resistant, and Sensitive).}

\item{save_output}{Logical. If \code{TRUE}, saves the plot as a JPEG file in the output directory (default: \code{FALSE}).}

\item{label_size}{Numeric. Controls the size of location labels on the map. Default: \code{2.5}.}

\item{map_data}{A list containing the shape file and longitude-latitude data for mapping.}

\item{sacle_piechart_size}{Numeric. Scales the maximum pie chart size. Default: \code{0.035}}

\item{time}{Optional. A list defining time periods for filtering the data.}
}
\value{
A list containing:
\itemize{
\item \code{Bar_Chart}: A bar chart of haplotype frequencies with percentages.
\item \code{Pie_Chart}: A pie chart map of haplotype proportions by location.
\item \code{Haplotype_Summary_Table}: A data frame summarizing the counts and percentages of haplotypes,
including an "Others" category for low-frequency haplotypes.
}
}
\description{
This function generates the proportion of unique haplotypes for a specific gene (e.g, PfCRT)
}
\examples{
haplotype_proportion(df = GRC_data,
                     drug_col = "Chloroquine",
                     gene_col = "PfCRT")

}
