% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summarise_IBS_Data.R
\name{ibs_data_dl}
\alias{ibs_data_dl}
\title{Summaries IBS data by the different locations}
\usage{
ibs_data_dl(
  melted_ibs_matrix,
  ibs_threshold = 0.75,
  map_data,
  label_size = 2.5,
  breaks = seq(0, 8, 2),
  curve_degree = 0.5,
  save_output = TRUE
)
}
\arguments{
\item{ibs_threshold}{A numeric value between 0 and 1 representing the IBS threshold for pair counts. Default is 0.75.}

\item{map_data}{A list containing the shape file and longitude-latitude data for mapping.}

\item{label_size}{Numeric. Controls the size of location labels on the map. Default: \code{2.5}.}

\item{breaks}{Used to set the breaks for the thickness and color of the lines. Default: \code{seq(0, 8, 2)}.}

\item{save_output}{Logical. If \code{TRUE}, saves the plot as a JPEG file in the output directory (default: \code{FALSE}).}

\item{melted_ibs_data}{A data frame containing the melted IBS data.
This data frame should include the columns \code{LS1}, \code{LS2}, and \code{value},
where value represents the IBS score between two locations.}
}
\value{
A data frame and connectivity maps showing the proportion of samples pairing between different locations.
\itemize{
\item \code{LS1, LS2}: The locations in each pair.
\item \code{mean_IBS}: The mean IBS score for the pair.
\item \code{median_IBS}: The median IBS score for the pair.
\item \code{TotalPairCount}: The total number of pairs between the two locations.
\item \code{<ibs_threshold>_PairCount}: The total number of pairs with an IBS score greater than or equal to the threshold.
\item \code{TotalPairCount_per}: The percentage of total pairs for each location pair.
\item \code{<ibs_threshold>_PairCount.per}: The percentage of total pairs that meet or exceed the IBS threshold.
\item \code{<ibs_threshold>_LocationPairCount.per}: The percentage of pairs for each location that meet or exceed the IBS threshold.
\item \code{Lat1, Long1, Lat2, Long2}: The latitude and longitude for both locations in each pair.
}
}
\description{
Summarize the IBS score by different location pairs.
}
\examples{
Example usage:
ibs_data_dl(melted_ibs_matrix = IBS_Melted_Matrix,
            map_data = geo_data,
            ibs_threshold = 0.75)

}
