% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summarise_IBS_Data.R
\name{ibs_data_sl}
\alias{ibs_data_sl}
\title{Summaries IBS data by the same locations.}
\usage{
ibs_data_sl(
  melted_ibs_matrix,
  ibs_threshold = 0.75,
  map_data,
  label_size = 2.5,
  circle_num_size = 3.1,
  scale_circle_size = 10,
  save_output = TRUE
)
}
\arguments{
\item{melted_ibs_matrix}{A data frame containing the melted IBS matrix}

\item{ibs_threshold}{A numeric value between 0 and 1 representing the IBS threshold for pair counts. Default is 0.75.}

\item{map_data}{A list containing the shape file and longitude-latitude data for mapping.}

\item{label_size}{Numeric. Controls the size of location labels on the map. Default: \code{2.5}.}

\item{circle_num_size}{Numeric. Controls the numbers inside the circles. Default: \code{3.1}.}

\item{scale_circle_size}{Numeric. Scales the maximum circle size. Default: \code{10}.}

\item{save_output}{Logical. If \code{TRUE}, saves the plot as a JPEG file in the output directory (default: \code{FALSE}).}
}
\value{
A data data frame and map showing the the median Ibs score for each location.
}
\description{
Summarize the IBS score by the same location pair. Getting the Median IBS per for each location.
}
\examples{
Example usage:
ibs_data_sl(melted_ibs_matrix = IBS_Melted_Matrix,
            map_data = geo_data,
            ibs_threshold = 0.75)

}
