% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summarise_IBS_Data.R
\name{ibs_data_summary}
\alias{ibs_data_summary}
\title{Manipulate the IBS Group data frames}
\usage{
ibs_data_summary(melted_ibs_data, ibs_th = 0.75)
}
\value{
A data frame summarizing the IBS scores between location pairs. The output data frame contains the following columns:
\itemize{
\item \code{LS1, LS2}: The locations in each pair.
\item \code{mean_IBS}: The mean IBS score for the pair.
\item \code{median_IBS}: The median IBS score for the pair.
\item \code{TotalPairCount}: The total number of pairs between the two locations.
\item \code{<ibs_threshold>_PairCount}: The number of pairs with an IBS score greater than or equal to the threshold.
\item \code{TotalPairCount_per}: The percentage of total pairs in the data set.
\item \code{<ibs_threshold>_PairCountGroup.per}: The percentage of total pairs that meet or exceed the IBS threshold.
\item \code{<ibs_threshold>_LocationPairCount.per}: The percentage of pairs for each location that meet or exceed the IBS threshold.
\item \code{Lat1, Long1, Lat2, Long2}: The latitude and longitude for both locations in each pair.
}
}
\description{
This function processes and summarizes a melted IBS matrix
It calculates the mean and median IBS values for each location pair, counts the total number of pairs,
and determines the number of pairs above a user-defined IBS threshold.
The function also calculates various percentages and appends geographical coordinates for each location.
}
\keyword{internal}
