% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBS_Heat_Maps.R
\name{ibs_heat_map}
\alias{ibs_heat_map}
\title{Generate Heat Maps using the IBS matrix}
\usage{
ibs_heat_map(
  df,
  snp_data,
  ibs_matrix,
  map_data,
  save_output = FALSE,
  drug_col = NULL
)
}
\arguments{
\item{df}{Final GRC data frame.}

\item{snp_data}{A data frame of SNP data, where each row represents a sample and each column corresponds to a
genetic locus (e.g., "Pf3D7_"). The row names should correspond to the "Sample Internal ID".}

\item{ibs_matrix}{Matrix of IBS scores for all sample pairs. Row and column names should match \verb{Sample Internal ID}.}

\item{map_data}{A list containing the shape file and longitude-latitude data for mapping.}

\item{save_output}{Logical. If \code{TRUE}, saves the plot as a JPEG file in the output directory (default: \code{FALSE}).}

\item{drug_col}{The name of the column representing the drug conditions (e.g., "Chloroquine" with
categories like Resistant, Mixed Resistant, and Sensitive).}
}
\value{
A list containing the heat maps.
}
\description{
This function generates a heat map representing Identity-by-State (IBS) scores between pairs of samples,
with colors assigned to each location dynamically. If \code{drug_col} is specified, heat maps are generated each containing
sample pairs associated with the category in the drug column.
}
\examples{
IBS_Heat_Map(df = GRC_Data,
             snp_data = barcode_data,
             map_data = geo_data,
             ibs_matrix = Ibs_matrix)

}
