% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mutation_Frequency_Plots.R
\name{Mutation_Frequency}
\alias{Mutation_Frequency}
\title{Mutation Frequency Calculation and Visualization}
\usage{
Mutation_Frequency(
  df,
  gene,
  col_gene,
  LongLat_data,
  shapeFile,
  col_location = NULL,
  location = NULL,
  year = NULL,
  include_mixed = FALSE
)
}
\arguments{
\item{df}{Final GRC dataframe}

\item{gene}{The gene to analyze ("pfcrt", "pfdhps", "pfdhfr", "pfmdr1")}

\item{col_gene}{The column in the df that contains haplotype data (e.g., "PfCRT").}

\item{LongLat_data}{Dataframe containing the locations and their geographical coo-ordinates, longitude and latitude}

\item{col_location}{Column in the dataframe that contains location information}

\item{location}{Location to filter by (optional)}

\item{year}{Year to filter by (optional)}

\item{include_mixed}{Boolean, whether to include mixed haplotypes in counts}

\item{shapFile}{An \code{sf} object representing the shape file of the country}
}
\value{
A list containing result tables and mutation plots
}
\description{
This function calculates the frequency of wild, mutated, mixed, and missing haplotypes for a specified gene,
and optionally filters the data by location and year. It also generates two tables: one for mutation types
by position and another for mutations by location. The function provides the option to include mixed haplotypes
in the counts for both tables.
}
