% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mutation_Frequency_Plots.R
\name{Mutation_Frequency}
\alias{Mutation_Frequency}
\title{Mutation Frequency Calculation and Visualization}
\usage{
Mutation_Frequency(
  df,
  gene,
  gene_col,
  drug_col,
  saveOutput = TRUE,
  period_name = "Full",
  time = NULL,
  labelSize = 2.5,
  circleNumSize = 3.1,
  scaleCircleSize = 10,
  include_mixed = FALSE,
  ...
)
}
\arguments{
\item{df}{Final GRC dataframe}

\item{gene}{The gene to analyze ("pfcrt", "pfdhps", "pfdhfr", "pfmdr1")}

\item{gene_col}{The column in the df that contains haplotype data (e.g., "PfCRT").}

\item{time}{Optional. A list defining time periods.}

\item{labelSize}{Used to set the size of the labels on the map.}

\item{circleNumSize}{Used to set th sizes of the numbers in th circles.}

\item{scaleCircleSize}{Used to scale the size of th circles.}

\item{include_mixed}{Boolean, whether to include mixed haplotypes in counts}
}
\description{
This function calculates the frequency of wild, mutated, mixed, and missing haplotypes for a specified gene.
It also generates two tables: one for mutation types by position and another for mutations by location.
The function provides the option to include mixed haplotypes in the counts for both tables.
The first Table is didspolayed in a form of a barshart with the percentages of the, wild, muataed, mixed and missing for the haplotype.
The second tabke is displayed using the maps with this percentages in circles and circle represewntaimg a location.
}
\examples{
Mutation_Frequency( GRC_Data, gene = pfcrt, gene_col = "PfCRT", drug_col = "Chloroquine")

}
