% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mutation_Frequency_Plots.R
\name{mutation_frequency}
\alias{mutation_frequency}
\title{Mutation Frequency Calculation and Visualization}
\usage{
mutation_frequency(
  df,
  gene,
  gene_col,
  save_output = TRUE,
  period_name = "Full",
  time = NULL,
  map_data,
  label_size = 2.5,
  circle_num_size = 3.1,
  scale_circle_size = 10,
  include_mixed = FALSE,
  ...
)
}
\arguments{
\item{df}{Final GRC data frame}

\item{gene}{The gene to analyze ("pfcrt", "pfdhps", "pfdhfr", "pfmdr1")}

\item{gene_col}{The column containing haplotype data (e.g., "PfCRT")}

\item{save_output}{Logical. If \code{TRUE}, saves the plot as a JPEG file in the output directory (default: \code{FALSE}).}

\item{time}{Optional. A list defining time periods.}

\item{map_data}{A list containing the shape file and longitude-latitude data for mapping.}

\item{label_size}{Numeric. Controls the size of location labels on the map. Default: \code{2.5}.}

\item{circle_num_size}{Numeric. Controls the numbers inside the circles. Default: \code{3.1}.}

\item{scale_circle_size}{Numeric. Scales the maximum circle size. Default: \code{10}.}

\item{include_mixed}{Boolean, whether to include mixed haplotypes in counts}
}
\description{
This function calculates the frequency of wild, mutated, mixed, and missing haplotypes for a specified gene.
It also generates two tables: one for mutation types by position and another for mutations by location.
The function provides the option to include mixed haplotypes in the counts for both tables.
The first plot is a bar chart with the percentages of, wild, mutated, mixed and missing for the haplotype.
The second plots are maps which show these percentages in circles for all available locations.
}
