% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mutation_Frequency_Plots.R
\name{mutation_frequency}
\alias{mutation_frequency}
\title{Mutation Frequency Calculation and Visualization}
\usage{
mutation_frequency(
  df,
  gene,
  gene_col,
  drug_col,
  save_output = TRUE,
  period_name = "Full",
  time = NULL,
  map_data,
  label_size = 2.5,
  circle_num_size = 3.1,
  scale_circle_size = 10,
  include_mixed = FALSE,
  ...
)
}
\arguments{
\item{df}{Final GRC dataframe}

\item{gene}{The gene to analyze ("pfcrt", "pfdhps", "pfdhfr", "pfmdr1")}

\item{gene_col}{The column in the df that contains haplotype data (e.g., "PfCRT").}

\item{drug_col}{The name of the column representing the drug conditions (e.g., "Chloroquine" with conditions, Resistant, mixed resistant and sensitive).}

\item{save_output}{Logical. Whether to save the output plots to files (default is FALSE).}

\item{time}{Optional. A list defining time periods.}

\item{map_data}{The metatdata list that contains your shapefile and Longitude Latitude data.}

\item{label_size}{Used to set the size of the labels on the map.}

\item{circle_num_size}{Used to set th sizes of the numbers in th circles.}

\item{scale_circle_size}{Used to scale the size of th circles.}

\item{include_mixed}{Boolean, whether to include mixed haplotypes in counts}
}
\description{
This function calculates the frequency of wild, mutated, mixed, and missing haplotypes for a specified gene.
It also generates two tables: one for mutation types by position and another for mutations by location.
The function provides the option to include mixed haplotypes in the counts for both tables.
The first plot is a bar chart with the percentages of, wild, mutated, mixed and missing for the haplotype.
The second plots are maps which show these percentages in circles and the circle representing a location.
}
\examples{
mutation_frequency( GRC_Data, gene = pfcrt, gene_col = "PfCRT", drug_col = "Chloroquine")

}
