% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCoA_Plots.R
\name{pcoa_plots}
\alias{pcoa_plots}
\title{Generate PCoA Plots using the IBS matrix}
\usage{
pcoa_plots(
  ibs_matrix,
  df,
  circle_size = 4,
  save_output = FALSE,
  drug_col = NULL
)
}
\arguments{
\item{ibs_matrix}{Matrix of IBS scores for all sample pairs. Row and column names should match \verb{Sample Internal ID}.}

\item{df}{Final GRC data frame}

\item{circle_size}{Numeric. Scales the circle size. Default: \code{4}.}

\item{save_output}{Logical. If \code{TRUE}, saves the plot as a JPEG file in the output directory (default: \code{FALSE}).}

\item{drug_col}{The name of the column representing the drug conditions (e.g., "Chloroquine" with
categories like Resistant, Mixed Resistant, and Sensitive).}
}
\value{
A list containing the pairwise pcoa scatter plots, (PCoA1 vs PCoA2,PCoA1 vs PCoA3, and PCoA2 vs PCoA3).
}
\description{
The \code{pcoa_plots} function performs Principal Coordinate Analysis (PCoA) on a given IBS matrix,
merging sample metadata and generating pairwise PCoA plots.
}
\details{
The function calculates three PCoA axes and produces pairwise scatter plots for these axes (PCoA1 vs PCoA2,
PCoA1 vs PCoA3, and PCoA2 vs PCoA3). If \code{drug_col} is specified, the resulting plots are facet wrapped by
the categories in the \code{drug_col}.
}
\examples{
pcoa_plots(ibs_matrix = Ibs_Matrix,
           df = GRC Data,
           drug_col = "Chloroquine")


}
