% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Drug_Condition_Proportion_Map.R
\name{Proportion_Map}
\alias{Proportion_Map}
\title{Proportion Map}
\usage{
Proportion_Map(
  df,
  drug_col,
  saveOutput = TRUE,
  period_name = "Full",
  time = NULL,
  labelSize = 2.5,
  circleNumSize = 3.1,
  scaleCircleSize = 10,
  ...
)
}
\arguments{
\item{df}{Final GRC dataframe.}

\item{drug_col}{The name of the column representing the drug conditions (e.g., "Chloroquine").}

\item{saveOutput}{Logical. Whether to save the output plots to files (default is FALSE).}

\item{time}{Optional. A list defining time periods.}

\item{labelSize}{Used to set the size of the labels on the map.}

\item{circleNumSize}{Used to set th sizes of the numbers in th circles.}

\item{scaleCircleSize}{Used to scale the size of th circles.}
}
\description{
This function generates maps displaying the percentages of conditions
(All.Resistant, Mixed.Resistant, Resistant, Sensitive and All_Resistant) by location
for the selected drug column. Each circle houses the percentage of the condition for that particular location
and is labeled using the name and samplecount for that location.
}
\examples{
Proportion_Map( GRC_Data, drug_col = "Chloroquine")

}
