% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleCountMap.R
\name{sample_count_map}
\alias{sample_count_map}
\title{Sample Count Map}
\usage{
sample_count_map(
  df,
  map_data,
  breaks = NULL,
  save_output = TRUE,
  period_name = "Full",
  label_size = 2.5,
  scale_circle_size = 11,
  time = NULL,
  ...
)
}
\arguments{
\item{df}{Final GRC data frame.}

\item{map_data}{A list containing the shape file and longitude-latitude data for mapping.}

\item{breaks}{Numeric vector. Sets the breaks for circle sizes. Default: \code{c(10, 50, 100, 200, 300, 400, 500)}.}

\item{save_output}{Logical. If \code{TRUE}, saves the plot as a JPEG file in the output directory (default: \code{FALSE}).}

\item{label_size}{Numeric. Controls the size of location labels on the map. Default: \code{2.5}.}

\item{scale_circle_size}{Numeric. Scales the maximum circle size. Default: \code{11}.}

\item{time}{Optional. A list defining time periods for filtering the data. Each element contains:
\itemize{
\item \code{type}: Either "year" or "period" to define time scope.
\item \code{start}: The start year of the time period.
\item \code{end}: The end year of the time period (for periods only).
\item \code{name}: The label to use for the period.
}}
}
\value{
A list containing the generated map \code{Sample_Count_Map} and a summary table \code{Sample_Count_Table}.
}
\description{
Creates a geographic map showing sample counts for each location based on longitude and latitude.
The map uses circle sizes to represent sample counts, with location labels displaying the counts in parentheses.
}
\examples{
# Example with time periods
time_periods <- list(list(type = "year", start = 2010, name = "2010"),
                     list(type = "period", start = 2015, end = 2019, name = "2015-2019"))

sample_count_map(df = GRC_data,
                 map_data = mapping_data,
                 time = time_periods)

}
