% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleCountMap.R
\name{SampleCountMap}
\alias{SampleCountMap}
\title{Sample Count Map}
\usage{
SampleCountMap(
  df,
  drug_col,
  mData,
  breaks = NULL,
  save_output = TRUE,
  period_name = "Full",
  label_size = 2.5,
  scale_circle_size = 11,
  time = NULL,
  ...
)
}
\arguments{
\item{df}{Final GRC dataframe.}

\item{drug_col}{The name of the column representing the drug conditions (e.g., "Chloroquine" with conditions, Resisatant, mixed resistant and sensitive).}

\item{mData}{The metatdata list that contains your shapefile and Longitude Latitude data.}

\item{breaks}{Used to set the breaks for the circle sizes}

\item{label_size}{Used to set the size of th labels on the map}

\item{scale_circle_size}{Used to scale the circle sizes.}

\item{time}{Optional. A list defining time periods, where each list element contains:
\itemize{
\item \code{type}: Either "year" or "period" to define time scope.
\item \code{start}: The start year of the time period.
\item \code{end}: The end year of the time period (for periods only).
\item \code{name}: The label to use for the period.
}

Example for generating time-period-based summary
time_periods <- list(list(type = "year", start = 2010, name = "2010"),
list(type = "period", start = 2015, end = 2019, name = "2015-2019"))}

\item{save_uutput}{Logical. Whether to save the output plots to files (default is FALSE).}
}
\description{
This function generates a map displaying the sample counts for each location based on the longitude and latitude coordinates.
It creates a map with circle for each location whose sizes are determined by how many samples come from that location.
The labels of the locations are also repelled from each circle with their sample counts in curly brackets.
}
\examples{
# Ensure that df, shapeFile, and LongLat_data are properly prepared in your global environment
SampleCountMap(GRC_data, drug_col = "Chlroquine",)

}
