% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleCountMap.R
\name{SampleCountMap}
\alias{SampleCountMap}
\title{Sample Count Map}
\usage{
SampleCountMap(summaryData, shapeFile, breaks = NULL)
}
\arguments{
\item{summaryData}{A dataframe containing location names, total sample counts, longitude (\code{long}), and latitude (\code{lat}) columns. It also includes other drug status information.}

\item{shapeFile}{An \code{sf} object representing the shape file of the country}

\item{breaks}{A numeric vector of break points for symbol sizes. If \code{NULL}, default breaks of c(10, 100, 200, 300) are used.}
}
\value{
The function saves a sample count map (\code{SampleCountMap.jpeg}) in the \code{savePath} directory and does not return any object to the R console.
}
\description{
This function generates a map displaying the total sample counts for each location based on longitude and latitude coordinates.
It creates a proportional symbol map with sample counts represented by varying symbol sizes, and it labels each location with the total sample count.
The map is saved as a JPEG file in the \code{savePath} directory.
}
\details{
This function first converts the \code{summaryData} into an \code{sf} object using the longitude and latitude coordinates. It then creates a proportional symbol map, where symbol sizes represent total sample counts at each location.
Labels are added to each point with the location name and the total sample count. The map is saved as a JPEG file in the folder specified by \code{savePath}.
}
\examples{
# Ensure that summaryData and shapeFile are properly prepared and are in your global environment
SampleCountMap(summaryData, shapeFile)

}
